Split gmregclear at its original translation-unit boundaries
The retail .sdata2 pool for this range holds four copies of the signed
int-to-float magic, three of the unsigned one, and duplicate 100.0f, 0.5f,
1.0f and 2.0f. A single MWCC unit pools each constant once, so the range
was several units. Two of the boundaries are recoverable exactly:

  - gm_GetAdventureData at 0x8017E424 opens a unit owning the adventure
    stage table (.data 0x803D7AC0) and UnkAdventureData (.bss 0x80472C30);
  - gm_GetAllStarData at 0x8017EB30 opens one owning the allstar stage
    table (.data 0x803D85F0) and its state (.bss 0x80472CB0);
  - 0x8017F008 opens gmregclear proper, whose .data starts at
    lbl_803D8B88 -- the string-pool base that fn_8017F2A4, fn_801803FC
    and fn_80180630 anchor on.

Every section divides along those lines without a remainder: .rodata,
.bss, .sdata, .sbss and .sdata2 each partition by which unit's functions
reference them, and the 4- and 2-byte holes before lbl_803D85F0 and
lbl_803D8B88 fall out as linker alignment rather than needing padding
members.

gm_17E4 and gm_17EB match completely, functions and data, and are marked
Matching. gm_17C0 and gmregclear are byte-exact in every data section
they own except gmregclear's .sdata2, which still merges the two units
remaining in its tail. fn_80180630 improves from 97.53% to 98.28% now
that it anchors on the right base.
