+++ config/GALE01/symbols.txt
@@ -31538,10 +31538,10 @@ lbl_804DA430 = .sdata2:0x804DA430; // type:object size:0x8 align:8 data:double
 lbl_804DA438 = .sdata2:0x804DA438; // type:object size:0x4 align:4 data:float
 lbl_804DA43C = .sdata2:0x804DA43C; // type:object size:0x4 align:4 data:float
 lbl_804DA440 = .sdata2:0x804DA440; // type:object size:0x4 align:4 data:float
-lbl_804DA448 = .sdata2:0x804DA448; // type:object size:0x4 align:4 data:float
-lbl_804DA450 = .sdata2:0x804DA450; // type:object size:0x8 align:8 data:double
-lbl_804DA458 = .sdata2:0x804DA458; // type:object size:0x4 align:4 data:float
-lbl_804DA460 = .sdata2:0x804DA460; // type:object size:0x8 align:8 data:double
+@217 = .sdata2:0x804DA448; // type:object size:0x4 scope:local align:4 data:float
+@219 = .sdata2:0x804DA450; // type:object size:0x8 scope:local align:8 data:double
+@197 = .sdata2:0x804DA458; // type:object size:0x4 scope:local align:4 data:float
+@199 = .sdata2:0x804DA460; // type:object size:0x8 scope:local align:8 data:double
 lbl_804DA468 = .sdata2:0x804DA468; // type:object size:0x4 align:4 data:float
 lbl_804DA46C = .sdata2:0x804DA46C; // type:object size:0x4 align:4 data:float
 lbl_804DA470 = .sdata2:0x804DA470; // type:object size:0x4 align:4 data:float
