+++ configure.py
@@ -1152,6 +1152,9 @@ config.libs = [
             Object(Matching, "melee/gm/gm_17AD.c"),
             Object(Matching, "melee/gm/gm_17BA.c"),
             Object(Matching, "melee/gm/gmregcommon.c"),
+            Object(Linkable, "melee/gm/gm_17C0.c"),
+            Object(Matching, "melee/gm/gm_17E4.c"),
+            Object(Matching, "melee/gm/gm_17EB.c"),
             Object(Linkable, "melee/gm/gmregclear.c"),
             Object(Linkable, "melee/gm/gm_1832.c"),
             Object(Linkable, "melee/gm/gm_1884.c"),
