+++ src/melee/gm/gm_1736.c
@@ -1,9 +1,10 @@
 #include "gm_1601.h"
 #include "gm_16F1.h"
+#include "gm_17C0.h"
+#include "gm_17E4.h"
 #include "gm_1A3F.h"
 #include "gmevent.h"
 #include "gmmain_lib.h"
-#include "gmregclear.h"
 #include "types.h"
 
 #include "ft/forward.h"
