+++ src/melee/gm/gm_17E4.h
@@ -1,0 +1,30 @@
+#ifndef MELEE_GM_17E4_H
+#define MELEE_GM_17E4_H
+
+#include <placeholder.h>
+#include <platform.h>
+
+#include <melee/gm/forward.h>
+
+/* 17E424 */ UnkAdventureData* gm_GetAdventureData(void);
+/* 17E430 */ u8 gm_8017E430(void); ///< returns a player slot
+/* 17E440 */ u8 gm_8017E440(void);
+/* 17E48C */ u8 gm_8017E48C(GameModeState*);
+/* 17E4C4 */ struct gm_803DE650_t* gm_8017E4C4(u8);
+/* 17E500 */ u8 gm_8017E500(u8, u8);
+/* 17E528 */ f32 gm_8017E528(u8, u8);
+/* 17E578 */ f32 gm_8017E578(u8, u8);
+/* 17E5C8 */ u8 gm_8017E5C8(u8, u8, u8);
+/* 17E5FC */ u8 gm_8017E5FC(u8, u8, u8);
+/* 17E630 */ u8 gm_8017E630(u8 difficulty, u8 stage_slot, u8 arg2);
+/* 17E664 */ f32 gm_8017E664(u8, u8);
+/* 17E6B4 */ f32 gm_8017E6B4(u8, u8);
+/* 17E704 */ u8 gm_8017E704(u8 difficulty, u8 stage_slot, u8 arg2);
+/* 17E738 */ u8 gm_8017E738(u8 difficulty, u8 stage_slot, u8 arg2);
+/* 17E76C */ u8 gm_8017E76C(u8 difficulty, u8 stage_slot, u8 arg2);
+/* 17E7A0 */ void gm_8017E7A0(u8);
+/* 17E7E0 */ bool gm_8017E7E0(void);
+/* 17E7FC */ void gm_8017E7FC(u8);
+/* 17E8A4 */ void fn_8017E8A4(int);
+
+#endif
