+++ src/melee/gm/gm_17EB.h
@@ -1,0 +1,24 @@
+#ifndef MELEE_GM_17EB_H
+#define MELEE_GM_17EB_H
+
+#include <placeholder.h>
+#include <platform.h>
+
+#include <melee/gm/forward.h>
+
+/* 17EB30 */ UnkAllstarData* gm_GetAllStarData(void);
+/* 17EB3C */ u8 gm_8017EB3C(u8, u8);
+/* 17EB64 */ u8 gm_8017EB64(u8 difficulty, u8 stage_slot, u8 arg2);
+/* 17EB98 */ u8 gm_8017EB98(u8 difficulty, u8 stage_slot, u8 arg2);
+/* 17EBCC */ u8 gm_8017EBCC(u8 difficulty, u8 stage_slot, u8 arg2);
+/* 17EC00 */ f32 gm_8017EC00(u8, u8);
+/* 17EC50 */ f32 gm_8017EC50(u8, u8);
+/* 17ECA0 */ u8 gm_8017ECA0(u8 difficulty, u8 stage_slot, u8 arg2);
+/* 17ECD4 */ u8 gm_8017ECD4(u8 difficulty, u8 stage_slot, u8 arg2);
+/* 17ED08 */ u8 gm_8017ED08(u8 difficulty, u8 stage_slot, u8 arg2);
+/* 17ED3C */ f32 gm_8017ED3C(u8, u8);
+/* 17ED8C */ f32 gm_8017ED8C(u8, u8);
+/* 17EDDC */ bool fn_8017EDDC(void);
+/* 17EE40 */ void fn_8017EE40(int);
+
+#endif
