+++ src/melee/gm/gm_1A9B.c
@@ -16,10 +16,10 @@
 
 #include "gm/forward.h"
 
+#include "gm/gm_17C0.h"
 #include "gm/gm_1A36.h"
 #include "gm/gm_1A3F.h"
 #include "gm/gmevent.h"
-#include "gm/gmregclear.h"
 #include "gm/gmregtyfall.h"
 #include "lb/lbaudio_ax.h"
 #include "lb/lbbgflash.h"
