+++ src/melee/gr/grcorneria.c
@@ -16,9 +16,9 @@
 
 #include "cm/camera.h"
 #include "ft/ftlib.h"
+#include "gm/gm_17E4.h"
 #include "gm/gm_1832.h"
 #include "gm/gm_1A45.h"
-#include "gm/gmregclear.h"
 #include "if/ifcoget.h"
 #include "if/ifstatus.h"
 #include "it/item.h"
